Add files via upload
We will get to know about lightning and heat waves using live weather APIs and warnings issued/predicted by the meteorological department. Updates regarding the calamity will be given in intervals. Users will be given a choice to include a temporary location in case of them being away from home. Messages /alerts are sent to the people with default location from UIDAI database and temporary location in bulk with alerts of particular regions. A database of alerts/warnings will also be maintained to study the weather pattern and to prepare beforehand.

--------------------------------------------------------------------------------------------------WORKING-----------------------------------------------------------------------------------
The Open Weather API is used to get the atmospheric weather data of the specified location. Our script filters though the 2-day early forecast and the hourly check. It gives priority to the hourly check. If the hour wise check returns a abnormal weather, SMS API is activated. It sends a formatted SMS to the public in that particular location (Phone numbers of the public can be obtained securely from the UIDAI along with the location). Along with the alerts, instructions are sent with time intervals unique to the calamity. The SMS is sent until the calamity comes to an end. If the hourly check returns a normal weather data, the 2-day early check is processed. Similar process is followed with the SMS API. Default location of a person will be set according to the UIDAI database, but it can be changed through the weekly prompt.